🧩 P8N5XR task: record opened PR
